fix: fix package structure for ast and kit packages
