+++ VERSION
@@ -1,1 +1,1 @@
-2.0.0-next.188
+2.0.0-beta.187
