+++ packages/plugins/eslint-plugin-react-debug/package.json
@@ -1,6 +1,6 @@
 {
   "name": "eslint-plugin-react-debug",
-  "version": "2.0.0-next.188",
+  "version": "2.0.0-beta.187",
   "description": "ESLint React's ESLint plugin for debugging related rules.",
   "keywords": [
     "react",
