+++ packages/plugins/eslint-plugin-react-dom/package.json
@@ -1,6 +1,6 @@
 {
   "name": "eslint-plugin-react-dom",
-  "version": "2.0.0-next.188",
+  "version": "2.0.0-beta.187",
   "description": "ESLint React's ESLint plugin for React DOM related rules.",
   "keywords": [
     "react",
