+++ packages/plugins/eslint-plugin-react-naming-convention/package.json
@@ -1,6 +1,6 @@
 {
   "name": "eslint-plugin-react-naming-convention",
-  "version": "2.0.0-next.188",
+  "version": "2.0.0-beta.187",
   "description": "ESLint React's ESLint plugin for naming convention related rules.",
   "keywords": [
     "react",
