+++ packages/plugins/eslint-plugin-react-x/package.json
@@ -1,6 +1,6 @@
 {
   "name": "eslint-plugin-react-x",
-  "version": "2.0.0-next.188",
+  "version": "2.0.0-beta.187",
   "description": "A set of composable ESLint rules for for libraries and frameworks that use React as a UI runtime.",
   "keywords": [
     "react",
