+++ packages/plugins/eslint-plugin-react-x/src/rules/no-unnecessary-use-prefix.ts
@@ -1,6 +1,6 @@
 import * as AST from "@eslint-react/ast";
 import * as ER from "@eslint-react/core";
-import { type RuleContext, type RuleFeature, isViMockCallback } from "@eslint-react/kit";
+import { type RuleContext, type RuleFeature } from "@eslint-react/kit";
 import type { TSESTree } from "@typescript-eslint/types";
 import type { RuleListener } from "@typescript-eslint/utils/ts-eslint";
 import type { CamelCase } from "string-ts";
@@ -64,7 +64,7 @@ export function create(context: RuleContext<MessageID, []>): RuleListener {
           continue;
         }
         // Skip hooks that are in a vi mock callback
-        if (AST.findParentNode(node, isViMockCallback) != null) {
+        if (AST.findParentNode(node, AST.isViMockCallback) != null) {
           continue;
         }
         context.report({
