+++ packages/plugins/eslint-plugin-react-x/src/rules/no-unstable-default-props.ts
@@ -1,12 +1,7 @@
 import * as AST from "@eslint-react/ast";
 import * as ER from "@eslint-react/core";
 import { getOrElseUpdate } from "@eslint-react/eff";
-import {
-  type ObjectDestructuringVariableDeclarator,
-  type RuleContext,
-  type RuleFeature,
-  SEL_OBJECT_DESTRUCTURING_VARIABLE_DECLARATOR,
-} from "@eslint-react/kit";
+import { type RuleContext, type RuleFeature } from "@eslint-react/kit";
 import * as VAR from "@eslint-react/var";
 import { AST_NODE_TYPES as T } from "@typescript-eslint/types";
 import type { RuleListener } from "@typescript-eslint/utils/ts-eslint";
@@ -41,10 +36,19 @@ export default createRule<[], MessageID>({
 
 export function create(context: RuleContext<MessageID, []>): RuleListener {
   const { ctx, listeners } = ER.useComponentCollector(context);
-  const declarators = new WeakMap<AST.TSESTreeFunction, ObjectDestructuringVariableDeclarator[]>();
+  const declarators = new WeakMap<AST.TSESTreeFunction, AST.ObjectDestructuringVariableDeclarator[]>();
 
   return {
     ...listeners,
+    [AST.SEL_OBJECT_DESTRUCTURING_VARIABLE_DECLARATOR](node: AST.ObjectDestructuringVariableDeclarator) {
+      const functionEntry = ctx.getCurrentEntry();
+      if (functionEntry == null) return;
+      getOrElseUpdate(
+        declarators,
+        functionEntry.node,
+        () => [],
+      ).push(node);
+    },
     "Program:exit"(program) {
       const components = ctx.getAllComponents(program);
       for (const { node: component } of components.values()) {
@@ -89,15 +93,6 @@ export function create(context: RuleContext<MessageID, []>): RuleListener {
           });
         }
       }
-    },
-    [SEL_OBJECT_DESTRUCTURING_VARIABLE_DECLARATOR](node: ObjectDestructuringVariableDeclarator) {
-      const functionEntry = ctx.getCurrentEntry();
-      if (functionEntry == null) return;
-      getOrElseUpdate(
-        declarators,
-        functionEntry.node,
-        () => [],
-      ).push(node);
     },
   };
 }
