+++ packages/utilities/ast/src/index.ts
@@ -1,10 +1,27 @@
 export * from "./array";
+export * from "./array-index";
+export * from "./array-method";
 export * from "./class-id";
 export * from "./expression";
+export * from "./expression-base";
+export * from "./expression-is";
+export * from "./expression-nested";
 export * from "./function";
+export * from "./function-id";
+export * from "./function-init-path";
+export * from "./function-is";
 export * from "./literal";
 export * from "./misc";
 export * from "./node";
+export * from "./node-equal";
+export * from "./node-format";
+export * from "./node-is";
+export * from "./node-selectors";
+export type * from "./node-types";
+export * from "./process-env-node-env";
 export * from "./promise-then";
 export * from "./property-name";
 export * from "./traverse";
+export type * from "./traverse-down";
+export * from "./traverse-up";
+export * from "./vitest-mock";
