+++ packages/utilities/ast/src/process-env-node-env.ts
@@ -1,6 +1,6 @@
-import * as AST from "@eslint-react/ast";
 import type { unit } from "@eslint-react/eff";
 import { AST_NODE_TYPES as T, type TSESTree } from "@typescript-eslint/types";
+import { isLiteral } from "./literal";
 
 /**
  * Check if the given node is a member expression that accesses `process.env.NODE_ENV`
@@ -34,40 +34,11 @@ export function isProcessEnvNodeEnvCompare(
   if (node == null) return false;
   if (node.type !== T.BinaryExpression) return false;
   if (node.operator !== operator) return false;
-  if (isProcessEnvNodeEnv(node.left) && AST.isLiteral(node.right, "string")) {
+  if (isProcessEnvNodeEnv(node.left) && isLiteral(node.right, "string")) {
     return node.right.value === value;
   }
-  if (AST.isLiteral(node.left, "string") && isProcessEnvNodeEnv(node.right)) {
+  if (isLiteral(node.left, "string") && isProcessEnvNodeEnv(node.right)) {
     return node.left.value === value;
   }
   return false;
-}
-
-/**
- * Checks if the given node is a `vi.mock`.
- * @param node The node to check
- * @returns `true` if the node is a `vi.mock`, otherwise `false`.
- * @internal
- */
-export function isViMock(node: TSESTree.Node | null | unit): node is TSESTree.MemberExpression {
-  return node != null
-    && node.type === T.MemberExpression
-    && node.object.type === T.Identifier
-    && node.object.name === "vi"
-    && node.property.type === T.Identifier
-    && node.property.name === "mock";
-}
-
-/**
- * Checks if the given node is a `vi.mock` callback.
- * @param node The node to check
- * @returns `true` if the node is a `vi.mock` callback, otherwise `false`.
- * @internal
- */
-export function isViMockCallback(node: TSESTree.Node | null | unit) {
-  return node != null
-    && AST.isFunction(node)
-    && node.parent.type === T.CallExpression
-    && isViMock(node.parent.callee)
-    && node.parent.arguments[1] === node;
 }
