--- packages/utilities/kit/docs/functions/isProcessEnvNodeEnv.md
@@ -1,25 +1,0 @@
-[**@eslint-react/kit**](../README.md)
-
-***
-
-[@eslint-react/kit](../README.md) / isProcessEnvNodeEnv
-
-# Function: isProcessEnvNodeEnv()
-
-> **isProcessEnvNodeEnv**(`node`): `node is MemberExpression`
-
-Check if the given node is a member expression that accesses `process.env.NODE_ENV`
-
-## Parameters
-
-### node
-
-The AST node
-
-`undefined` | `null` | `Node`
-
-## Returns
-
-`node is MemberExpression`
-
-True if the node is a member expression that accesses `process.env.NODE_ENV`, false otherwise
