--- packages/utilities/kit/docs/type-aliases/DisplayNameAssignmentExpression.md
@@ -1,39 +1,0 @@
-[**@eslint-react/kit**](../README.md)
-
-***
-
-[@eslint-react/kit](../README.md) / DisplayNameAssignmentExpression
-
-# Type Alias: DisplayNameAssignmentExpression
-
-> **DisplayNameAssignmentExpression** = `TSESTree.AssignmentExpression` & `object`
-
-## Type Declaration
-
-### left
-
-> **left**: `TSESTree.MemberExpression` & `object`
-
-#### Type Declaration
-
-##### property
-
-> **property**: `TSESTree.Identifier` & `object`
-
-###### Type Declaration
-
-###### name
-
-> **name**: `"displayName"`
-
-### operator
-
-> **operator**: `"="`
-
-### right
-
-> **right**: `TSESTree.Literal`
-
-### type
-
-> **type**: `"AssignmentExpression"`
