--- packages/utilities/kit/docs/type-aliases/ImplicitReturnArrowFunctionExpression.md
@@ -1,15 +1,0 @@
-[**@eslint-react/kit**](../README.md)
-
-***
-
-[@eslint-react/kit](../README.md) / ImplicitReturnArrowFunctionExpression
-
-# Type Alias: ImplicitReturnArrowFunctionExpression
-
-> **ImplicitReturnArrowFunctionExpression** = `TSESTree.ArrowFunctionExpression` & `object`
-
-## Type Declaration
-
-### body
-
-> **body**: `TSESTree.Expression`
