--- packages/utilities/kit/docs/type-aliases/ObjectDestructuringVariableDeclarator.md
@@ -1,19 +1,0 @@
-[**@eslint-react/kit**](../README.md)
-
-***
-
-[@eslint-react/kit](../README.md) / ObjectDestructuringVariableDeclarator
-
-# Type Alias: ObjectDestructuringVariableDeclarator
-
-> **ObjectDestructuringVariableDeclarator** = `TSESTree.VariableDeclarator` & `object`
-
-## Type Declaration
-
-### id
-
-> **id**: `TSESTree.ObjectPattern`
-
-### init
-
-> **init**: `TSESTree.Identifier`
