--- packages/utilities/kit/docs/variables/SEL_DISPLAY_NAME_ASSIGNMENT_EXPRESSION.md
@@ -1,9 +1,0 @@
-[**@eslint-react/kit**](../README.md)
-
-***
-
-[@eslint-react/kit](../README.md) / SEL\_DISPLAY\_NAME\_ASSIGNMENT\_EXPRESSION
-
-# Variable: SEL\_DISPLAY\_NAME\_ASSIGNMENT\_EXPRESSION
-
-> `const` **SEL\_DISPLAY\_NAME\_ASSIGNMENT\_EXPRESSION**: `string`
