--- packages/utilities/kit/docs/variables/SEL_IMPLICIT_RETURN_ARROW_FUNCTION_EXPRESSION.md
@@ -1,9 +1,0 @@
-[**@eslint-react/kit**](../README.md)
-
-***
-
-[@eslint-react/kit](../README.md) / SEL\_IMPLICIT\_RETURN\_ARROW\_FUNCTION\_EXPRESSION
-
-# Variable: SEL\_IMPLICIT\_RETURN\_ARROW\_FUNCTION\_EXPRESSION
-
-> `const` **SEL\_IMPLICIT\_RETURN\_ARROW\_FUNCTION\_EXPRESSION**: `"ArrowFunctionExpression[body.type!='BlockStatement']"` = `"ArrowFunctionExpression[body.type!='BlockStatement']"`
