--- packages/utilities/kit/docs/variables/SEL_OBJECT_DESTRUCTURING_VARIABLE_DECLARATOR.md
@@ -1,9 +1,0 @@
-[**@eslint-react/kit**](../README.md)
-
-***
-
-[@eslint-react/kit](../README.md) / SEL\_OBJECT\_DESTRUCTURING\_VARIABLE\_DECLARATOR
-
-# Variable: SEL\_OBJECT\_DESTRUCTURING\_VARIABLE\_DECLARATOR
-
-> `const` **SEL\_OBJECT\_DESTRUCTURING\_VARIABLE\_DECLARATOR**: `string`
