+++ packages/utilities/kit/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@eslint-react/kit",
-  "version": "2.0.0-next.188",
+  "version": "2.0.0-beta.187",
   "description": "ESLint React's plugin kit for building plugins and rules.",
   "homepage": "https://github.com/Rel1cx/eslint-react",
   "bugs": {
@@ -34,7 +34,6 @@
     "publish": "pnpm run build && pnpm run lint:publish"
   },
   "dependencies": {
-    "@eslint-react/ast": "workspace:*",
     "@eslint-react/eff": "workspace:*",
     "@typescript-eslint/utils": "^8.44.0",
     "ts-pattern": "^5.8.0",
