--- packages/utilities/kit/src/ast/index.ts
@@ -1,2 +1,0 @@
-export * from "./env-checks";
-export * from "./selectors";
