+++ packages/utilities/kit/src/index.ts
@@ -1,4 +1,3 @@
-export * from "./ast";
 export * from "./constants";
 export type * from "./types";
 export * from "./utils";
