+++ packages/utilities/var/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@eslint-react/var",
-  "version": "2.0.0-next.188",
+  "version": "2.0.0-beta.187",
   "description": "ESLint React's TSESTree AST utility module for static analysis of variables.",
   "homepage": "https://github.com/Rel1cx/eslint-react",
   "bugs": {
